Beta2 (#98)
* Update utilities.py

* Update rawfunctions.py

* refining code

* updating modules directory

* Update rawfunctions.py

* Update rawfunctions.py

* Update rawfunctions.py

* Update utilities.py

* function naming

* .

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* minor

+++ tronx/modules/tools.py
@@ -59,7 +59,7 @@ def convert_c(celsius):
 
 
 @app.on_message(gen("wlink", allow = ["sudo", "channel"]))
-async def get_word_links(_, m: Message):
+async def wordlink_handler(_, m: Message):
 	links = []
 	links.clear()
 
@@ -68,14 +68,14 @@ async def get_word_links(_, m: Message):
 			return await app.send_edit(m, "Please give some text to search in chat ...")
 
 		else:
-			info = await app.get_history(m.chat.id)
+			messages = await app.get_history(m.chat.id)
 			query = m.text.split(None, 1)[1]
-			m = await app.send_edit(m, "Finding word in this chat . . .", mono=True)
-			for words in info:
-				if query in words.text:
-					links.append(words.link)
+			m = await app.send_edit(m, "Finding word in this chat . . .", text_type=["mono"])
+			for msg in messages:
+				if query in msg.text:
+					links.append(msg.link)
 
-			await app.send_edit(m, f"**FOUND LINKS FOR:** `{query}`\n\n" +"\n".join(links))
+			await app.send_edit(m, f"**FOUND LINKS FOR:** `{query}`\n\n" + "\n".join(links))
 	except Exception as e:
 		await app.error(m, e)
 
@@ -83,7 +83,7 @@ async def get_word_links(_, m: Message):
 
 
 @app.on_message(gen(["cur", "currency"], allow = ["sudo", "channel"]))
-async def evaluate(_, m: Message):
+async def currency_handler(_, m: Message):
 	if app.long(m) <= 3:
 		return await app.send_edit(m, f"Use | `{app.PREFIX}cur 100 USD INR` or `{app.PREFIX}currency 100 USD INR`")
 
@@ -102,9 +102,9 @@ async def evaluate(_, m: Message):
 
 
 @app.on_message(gen(["temp", "temperature"], allow = ["sudo", "channel"]))
-async def evaluate(_, m: Message):
+async def temperature_handler(_, m: Message):
 	if app.long(m) <= 2:
-		return await app.send_edit(m, "How To Use: [INSTANT VIEW](https://telegra.ph/HOW-TO-USE-04-11)", disable_web_page_preview=True)
+		return await app.send_edit(m, f"How To Use: `{app.MyPrefix()[0]}temp 10 c`", disable_web_page_preview=True)
 
 	temp1 = m.text.split(None, 2)[1]
 	temp2 = m.text.split(None, 2)[2]
@@ -126,15 +126,15 @@ async def evaluate(_, m: Message):
 
 
 @app.on_message(gen("json", allow = ["sudo", "channel", "forward"]))
-async def json_of_msg(_, m: Message):
+async def messagejson_handler(_, m: Message):
 	reply = m.reply_to_message
 
 	data = str(reply) if reply else str(m)
 
 	try:
-		await app.send_edit(m, data, mono=True)
+		await app.send_edit(m, data, text_type=["mono"])
 	except Exception: # message too long
-		m = await app.send_edit(m, "Sending file . . .", mono=True)
+		m = await app.send_edit(m, "Sending file . . .", text_type=["mono"])
 		await app.create_file(m, "json.txt", data)
 		if m.from_user.is_self:
 			await m.delete()
@@ -143,7 +143,7 @@ async def json_of_msg(_, m: Message):
 
 
 @app.on_message(gen("ulink", allow = ["sudo", "channel"]))
-async def get_inlinelinks(app, m: Message):
+async def urllink_handler(app, m: Message):
 	reply = m.reply_to_message
 	cat = []
 	dog = []
@@ -166,26 +166,26 @@ async def get_inlinelinks(app, m: Message):
 		if msg:
 			await app.send_edit(m, f"`{msg}`")
 		else:
-			await app.send_edit(m, "There are no links in this message . . .", mono=True)
+			await app.send_edit(m, "There are no links in this message . . .", text_type=["mono"])
 	else:
-		await app.send_edit(m, "Try this command on url button message to get info of the button . . .", mono=True)
+		await app.send_edit(m, "Try this command on url button message to get info of the button . . .", text_type=["mono"])
 
 
 
 
 @app.on_message(gen("mlink", allow = ["sudo", "channel"]))
-async def get_message_links(_, m: Message):
+async def messagelink_handler(_, m: Message):
 	reply = m.reply_to_message
 	message = reply if reply else m
 
-	m = await app.send_edit(m, "Generating message link . . .", mono=True)
+	m = await app.send_edit(m, "Generating message link . . .", text_type=["mono"])
 	await app.send_edit(m, message.link)
 
 
 
 
 @app.on_message(gen("saved", allow = ["sudo", "channel"]))
-async def save_to_cloud(_, m: Message):
+async def saved_handler(_, m: Message):
 	if m.from_user.is_self:
 		await m.delete()
 	await m.reply_to_message.copy("me")
@@ -194,26 +194,31 @@ async def save_to_cloud(_, m: Message):
 
 
 @app.on_message(gen(["fwd", "frwd"], allow = ["sudo", "channel"]))
-async def forward_msgs(_, m: Message):
+async def forward_handler(_, m: Message):
 	reply = m.reply_to_message
 	try:
 
 		if reply and app.long(m) == 1:
 			await reply.forward(m.chat.id)
+			delete = True
 
 		elif reply and app.long(m) > 1:
 			await reply.forward(m.command[1])
+			delete = True
 
 		elif not reply and app.long(m) == 1:
-			await m.forward(m.text if m.text else "None")
+			await m.forward(m.chat.id)
+			delete = True
 
 		elif not reply and app.long(m) > 1:
-			await app.send_edit(m, "Sir reply to yours or someone's message. to forward.", mono=True, delme=4)
+			await app.send_edit(m, "Sir reply to yours or someone's message. to forward.", text_type=["mono"], delme=4)
+			delete = False
 
 		else:
-			await app.send_edit(m, "Something went wrong, please try again later !", mono=True, delme=4)
+			await app.send_edit(m, "Something went wrong, please try again later !", text_type=["mono"], delme=4)
+			delete = False
 
-		if m.from_user.is_self:
+		if m.from_user.is_self and delete:
 			await m.delete()
 
 	except Exception as e:
@@ -223,9 +228,9 @@ async def forward_msgs(_, m: Message):
 
 
 @app.on_message(gen(["spt", "speed", "speedtest"], allow = ["sudo", "channel"]))
-async def speed_tests(app, m: Message):
+async def speedtest_handler(app, m: Message):
 	if app.long(m) == 1:
-		await app.send_edit(m, "Testing speed . . .", mono=True)
+		await app.send_edit(m, "Testing speed . . .", text_type=["mono"])
 		test = speedtest.Speedtest()
 		test.get_best_server()
 		test.download()
@@ -241,7 +246,7 @@ async def speed_tests(app, m: Message):
 		if teks:
 			await app.send_edit(m, teks)
 		else:
-			await app.send_edit(m, "Something went wrong !", mono=True, delme=5)
+			await app.send_edit(m, "Something went wrong !", text_type=["mono"], delme=5)
 	elif app.long(m) > 1 and "pic" in m.command[1]:
 		m = await app.send_edit(m, "Calculating Speed (pic) . . .")
 
@@ -270,14 +275,14 @@ async def speed_tests(app, m: Message):
 			)
 			await m.delete()
 		else:
-			await app.send_edit(m, "Something went wrong !", mono=True, delme=5)
+			await app.send_edit(m, "Something went wrong !", text_type=["mono"], delme=5)
 
 
 
 
 
 @app.on_message(gen(["cc", "cchats"], allow = ["sudo", "channel"]))
-async def common_chats(_, m):
+async def commonchat_handler(_, m):
 	try:
 		reply = m.reply_to_message
 		if reply:
@@ -292,7 +297,7 @@ async def common_chats(_, m):
 			else:
 				await app.send_edit(m, f"**Common chats with:** `{reply.from_user.first_name}`\n\n" + "`None`")
 		else:
-			await app.send_edit(m, "Please reply to someone . . .", mono=True, delme=4)
+			await app.send_edit(m, "Please reply to someone . . .", text_type=["mono"], delme=4)
 	except Exception as e:
 		await app.error(m, e)
